move "Look & Feel/Book details" to its own widget
